requirements.txt
beautifulsoup4==4.12.3
firebase_admin==6.4.0
fuzzywuzzy==0.18.0
networkx==3.2.1
pandas==2.2.1
python-dotenv==1.0.1
Requests==2.31.0
scikit_learn==1.4.1.post1
streamlit==1.30.0
